Added HAProxyLogRecord type to allow logging an HTTP request from the time it hits the port all the way through the pool down to the server.
First pass commit, still need to create a logRecord interface which HAProxy implements to allow log formats to be switched out easily. Also, need to write methods for HAProxyLog record which allow the ResponseWriter and Request variables to be private. Thus allowing the logrecord to keep track of all writes done to the response and all changes made to it. As of right now some data fields are not being filled in yet including all the wait times and the pid of the logger. Also, metrics needs to be generalized instead of having one for server and one for pool and port. Lastly, the actual log output needs to be tweaked some still to match up with the spec